add a protocol called UIPickerViewDataSource
set the ViewController.swift as the datasource for the picker. Find viewDidLoad() and add set the ViewController class as the datasource to the currencyPicker object. Provide the data and add the implementation for the first method numberOfComponents(in:) to determine how many columns we want in our picker. Added rows this picker should have using the pickerView:numberOfRowsInComponent: method. Use the count method on the currencyArray in the CoinManager to get that information. Update the PickerView with some titles and detect when it is interacted with. Set up the PickerView’s delegate methods.Add the protocol UIPickerViewDelegate to the class declaration and set the ViewController class as the delegate of the currencyPicker.