Added a functioning Picker View
Add the delegate method pickerView(titleForRow:) This method expects a String as an output. The String is the title for a given row. When the PickerView is loading up, it will ask its delegate for a row title and call the above method once for every row. So when it is trying to get the title for the first row, it will pass in a row value of 0 and a component (column) value of 0. Inside the method. Use the row Int to pick the title from our currencyArray.